Party selection in stateless view
