+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/utils/urlHelper.ts
@@ -164,3 +164,7 @@ export function getRulehandlerUrl(layoutset: string) {
 export function getCalculatePageOrderUrl() {
   return `${appPath}/instances/${altinnWindow.instanceId}/pages/order`;
 }
+
+export function getPartyValidationUrl(partyId: string) {
+  return `${appPath}/api/v1/parties/validateInstantiation?partyId=${partyId}`;
+}
